Revert "Extract QueryClientTestBase"
This reverts commit a41e2ec1352a1084d1a2e4e8c5386fd7d5116f14.
